+++ src/gui/Neuvisysthread.h
@@ -30,7 +30,7 @@ protected:
     cv::Mat m_leftEventDisplay;
     cv::Mat m_rightEventDisplay;
     std::map<size_t, cv::Mat> m_weightDisplay;
-    std::vector<std::reference_wrapper<const std::vector<long>>> m_spikeTrain;
+    std::vector<std::reference_wrapper<const std::vector<size_t>>> m_spikeTrain;
     std::vector<bool> m_motorDisplay;
     double m_eventRate{};
     size_t m_mode = 0;
@@ -88,8 +88,8 @@ signals:
     void displayStatistics(double event_rate, double on_off_ratio, double spike_rate, double threshold, double bias);
     void displayEvents(const cv::Mat &leftEventDisplay, const cv::Mat& rightEventDisplay);
     void displayWeights(const std::map<size_t, cv::Mat>& weightDisplay, size_t layer);
-    void displayPotential(double vreset, double threshold, const std::vector<std::pair<double, long>> &potentialTrain);
-    void displaySpike(const std::vector<std::reference_wrapper<const std::vector<long>>> &spikeTrain, double time);
+    void displayPotential(double vreset, double threshold, const std::vector<std::pair<double, size_t>> &potentialTrain);
+    void displaySpike(const std::vector<std::reference_wrapper<const std::vector<size_t>>> &spikeTrain, double time);
     void displayReward(const std::vector<double> &rewardTrain, const std::vector<double> &valueTrain, const std::vector<double> &valueDotTrain, const std::vector<double> &tdTrain);
     void displayAction(const std::vector<bool> &motorActivation);
     void networkConfiguration(const std::string &sharingType, const std::vector<std::vector<size_t>> &layerPatches, const std::vector<size_t> &layerSizes, const
